- force long for minutes

+++ src/iroffer_main.c
@@ -3049,7 +3049,7 @@ void sendxdccfile(const char* nick, const char* hostname, const char* hostmask, 
    else if (!man && gdata.nonewcons > gdata.curtime ) {
       ioutput(CALLTYPE_MULTI_MIDDLE,OUT_S|OUT_L|OUT_D,COLOR_YELLOW," (No New Cons): ");
       notice(nick, "** The Owner Has Requested That No New Connections Are Made In The Next %li Minutes%s%s",
-             (gdata.nonewcons-gdata.curtime+1)/60,
+             (long)((gdata.nonewcons-gdata.curtime+1)/60),
              gdata.nosendmsg?", ":"",
              gdata.nosendmsg?gdata.nosendmsg:"");
       }
